% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_sales_details.R
\name{as_sales_details}
\alias{as_sales_details}
\title{Assign an object to the `sales_details` class}
\usage{
as_sales_details(x)
}
\arguments{
\item{x}{an object containing summary sales data}
}
\description{
Assign an object to the `sales_details` class
}
