% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{display}
\alias{display}
\title{Create a generic function for displaying data}
\usage{
display(x, ...)
}
\description{
Create a generic function for displaying data
}
\note{
Use to make a pretty DT::datatable
}
