% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_ship_date}
\alias{estimate_ship_date}
\title{Estimate the ship date based on the order date}
\usage{
estimate_ship_date(order_date)
}
\arguments{
\item{order_date}{datetime of order (POSIXct/vector)}
}
\description{
Estimate the ship date based on the order date
}
\note{
Factors in 2pm cutoff and no shipping on weekends
}
