% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_sales_details.R
\name{is_sales_details}
\alias{is_sales_details}
\title{Check if an object is of the `sales_details` class}
\usage{
is_sales_details(x)
}
\arguments{
\item{x}{an R object to check}
}
\description{
Check if an object is of the `sales_details` class
}
