% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_date_range}
\alias{split_date_range}
\title{Divide a date-range into subsequences of a given size}
\usage{
split_date_range(start_date, end_date, n = 7)
}
\arguments{
\item{start_date}{First day of date range (Date/scalar)}

\item{end_date}{Last day of date range (Date/scalar)}
}
\description{
Divide a date-range into subsequences of a given size
}
