% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_sales_details.R
\name{summary.sales_details}
\alias{summary.sales_details}
\title{Summarize a sales_details object}
\usage{
\method{summary}{sales_details}(x, ...)
}
\arguments{
\item{x}{object of 'sales_details' class}

\item{...}{grouping dimensions}
}
\description{
Summarize a sales_details object
}
