% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_add_inventory.R
\name{sv_add_inventory}
\alias{sv_add_inventory}
\title{Add inventory}
\usage{
sv_add_inventory(...)
}
\arguments{
\item{...}{args passed on to \code{\link{sv_update_inventory}}}
}
\description{
Add inventory
}
