% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_api.R
\name{sv_api}
\alias{sv_api}
\title{Call a SkuVault API endpoint and iterate over all pages}
\usage{
sv_api(...)
}
\arguments{
\item{...}{additional arguments passed to the API}
}
\description{
Call a SkuVault API endpoint and iterate over all pages
}
