% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_api.R
\name{sv_api_call}
\alias{sv_api_call}
\title{Make a SkuVault API call}
\usage{
sv_api_call(path, ...)
}
\arguments{
\item{path}{path to api endpoint}

\item{...}{additional named parameters}
}
\description{
Make a SkuVault API call
}
