% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_cancel_order.R
\name{sv_cancel_order}
\alias{sv_cancel_order}
\title{Cancel an order (and auto-unpick items)}
\usage{
sv_cancel_order(saleid)
}
\arguments{
\item{saleid}{SaleId of order to pick}
}
\description{
Cancel an order (and auto-unpick items)
}
