% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_create_kits.R
\name{sv_create_kits}
\alias{sv_create_kits}
\title{Create a new kit in Skuvault}
\usage{
sv_create_kits(x)
}
\arguments{
\item{x}{data.frame containing new kits (columns: Sku, Title, KitLines)}
}
\description{
Create a new kit in Skuvault
}
