% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_create_listing.R
\name{sv_create_listing}
\alias{sv_create_listing}
\title{Generate a ChannelAdvisor Listing from SkuVault product data}
\usage{
sv_create_listing(x)
}
\arguments{
\item{x}{data from a call to \code{\link{sv_get_products}}}
}
\description{
Generate a ChannelAdvisor Listing from SkuVault product data
}
