% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_create_multipacks.R
\name{sv_create_multipacks}
\alias{sv_create_multipacks}
\title{Generate multipacks based on SkuVault product data}
\usage{
sv_create_multipacks(x, n = c(2, 4, 8))
}
\arguments{
\item{x}{data for a single product (non-kit) from a call to \code{\link{sv_get_products}}}

\item{n}{pack sizes (integer/vector)}
}
\description{
Generate multipacks based on SkuVault product data
}
