% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_create_product.R
\name{sv_create_product}
\alias{sv_create_product}
\title{Create a new product in SkuVault}
\usage{
sv_create_product(Sku, Brand, Classification, Supplier, Cost, ...)
}
\arguments{
\item{Sku}{Sku}

\item{Brand}{Brand}

\item{Classification}{Classification}

\item{Supplier}{Supplier}

\item{Cost}{Cost}
}
\description{
Create a new product in SkuVault
}
