% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_create_products.R
\name{sv_create_products}
\alias{sv_create_products}
\title{Create new products in Skuvault}
\usage{
sv_create_products(df)
}
\arguments{
\item{df}{data.frame containing new products (columns: Sku, Description)}
}
\description{
Create new products in Skuvault
}
