% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_alternate_skus.R
\name{sv_get_alternate_skus}
\alias{sv_get_alternate_skus}
\title{Get a list of alternate SKUs}
\usage{
sv_get_alternate_skus(data = NULL)
}
\arguments{
\item{data}{data from call to \code{\link{skuvaultr::sv_get_products}}}
}
\description{
Get a list of alternate SKUs
}
