% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_cost_history.R
\name{sv_get_cost_history}
\alias{sv_get_cost_history}
\title{Construct cost history}
\usage{
sv_get_cost_history(skus = "hon1010")
}
\arguments{
\item{skus}{skus (character/vector)}
}
\description{
Construct cost history
}
