% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_empty_bins.R
\name{sv_get_empty_bins}
\alias{sv_get_empty_bins}
\title{Find all inventory bins recently emptied}
\usage{
sv_get_empty_bins(start_date = Sys.Date() - 7, end_date = Sys.Date())
}
\arguments{
\item{start_date}{first day of data (Date/scalar)}

\item{end_date}{last day of data (Date/scalar)}
}
\description{
Find all inventory bins recently emptied
}
