% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_inventory_value.R
\name{sv_get_inventory_value}
\alias{sv_get_inventory_value}
\title{Get total warehouse inventory value}
\usage{
sv_get_inventory_value(skus = NULL)
}
\arguments{
\item{skus}{product skus (character/vector)}
}
\description{
Get total warehouse inventory value
}
