% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_kits.R
\name{sv_get_kit_details}
\alias{sv_get_kit_details}
\title{Get kit details}
\usage{
sv_get_kit_details(kits)
}
\arguments{
\item{kits}{kit data from a call to \code{\link{sv_get_kits}}}
}
\description{
Get kit details
}
