% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_kits.R
\name{sv_get_kit_weights}
\alias{sv_get_kit_weights}
\title{Get kit weights}
\usage{
sv_get_kit_weights(kits)
}
\arguments{
\item{kits}{kit data from a call to \code{\link{sv_get_kits}}}
}
\description{
Get kit weights
}
