% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_kits.R
\name{sv_get_kits}
\alias{sv_get_kits}
\title{Get kits}
\usage{
sv_get_kits(skus = NULL, include_details = TRUE)
}
\arguments{
\item{skus}{skus (character/vector)}

\item{include_details}{include kit details? (weight/brand)}
}
\description{
Get kits
}
