% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_out_of_stock.R
\name{sv_get_out_of_stock}
\alias{sv_get_out_of_stock}
\title{Find all out-of-stock (drop-shipped) products}
\usage{
sv_get_out_of_stock(drop_ship_only = TRUE)
}
\arguments{
\item{drop_ship_only}{restrict to drop-shipped products (logical/scalar)}
}
\description{
Find all out-of-stock (drop-shipped) products
}
