% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_picking_speed.R
\name{sv_get_picking_speed}
\alias{sv_get_picking_speed}
\title{Get picking speed (overall; user-specific)}
\usage{
sv_get_picking_speed(
  start_date = Sys.Date(),
  end_date = Sys.Date(),
  show_plot = TRUE
)
}
\arguments{
\item{start_date}{first day of data (Date/scalar)}

\item{end_date}{last day of data (Date/scalar)}
}
\description{
Get picking speed (overall; user-specific)
}
