% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_picklists.R
\name{sv_get_picklists}
\alias{sv_get_picklists}
\title{Get picklist given a set of SaleIds}
\usage{
sv_get_picklists(saleids, data = NULL)
}
\arguments{
\item{saleids}{saleid(s) (character/vector)}

\item{data}{data returned from 'sv_get_sales'}
}
\description{
Get picklist given a set of SaleIds
}
