% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_reorder_report.R
\name{sv_get_reorder_report}
\alias{sv_get_reorder_report}
\title{Create a product reorder report}
\usage{
sv_get_reorder_report(n_days = 30, include_last_season = TRUE)
}
\arguments{
\item{n_days}{Number of days in-stock to order}
}
\description{
Create a product reorder report
}
