% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_sales.R
\name{sv_get_sales}
\alias{sv_get_sales}
\title{Get sales}
\usage{
sv_get_sales(start_date = Sys.Date() - 1, end_date = Sys.Date() - 1)
}
\arguments{
\item{start_date}{first day of data (Date/scalar)}

\item{end_date}{last day of data (Date/scalar)}
}
\description{
Get sales
}
