% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_sales.R
\name{sv_get_sales_7day}
\alias{sv_get_sales_7day}
\title{Get sales (<= 7 days only)}
\usage{
sv_get_sales_7day(
  start_date = Sys.Date() - 1,
  end_date = Sys.Date() - 1,
  order_id = NULL
)
}
\arguments{
\item{start_date}{first day of data (Date/scalar)}

\item{end_date}{last day of data (Date/scalar)}

\item{order_id}{order/sale ids (character/vector)}
}
\description{
Get sales (<= 7 days only)
}
