% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_sales_details.R
\name{sv_get_sales_details}
\alias{sv_get_sales_details}
\title{Get cleaned multi-channel sales data (1 row / sku)}
\usage{
sv_get_sales_details(...)
}
\arguments{
\item{...}{arguments passed on to \code{\link{sv_get_sales}}}
}
\description{
Get cleaned multi-channel sales data (1 row / sku)
}
