% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_tokens.R
\name{sv_get_tokens}
\alias{sv_get_tokens}
\title{Get a SkuVault API token set (TenantToken/UserToken)}
\usage{
sv_get_tokens(
  email = Sys.getenv("SKU_VAULT_EMAIL"),
  password = Sys.getenv("SKU_VAULT_PASSWORD")
)
}
\arguments{
\item{email}{SkuVault account email address}

\item{password}{SkuVault account password}
}
\description{
Get a SkuVault API token set (TenantToken/UserToken)
}
