% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_transactions.R
\name{sv_get_transactions}
\alias{sv_get_transactions}
\title{Get transactions}
\usage{
sv_get_transactions(start_date = Sys.Date() - 1, end_date = Sys.Date()
  - 1)
}
\description{
Get transactions
}
