% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_transactions.R
\name{sv_get_transactions_7day}
\alias{sv_get_transactions_7day}
\title{Get transactions (<= 7 days only)}
\usage{
sv_get_transactions_7day(start_date = Sys.Date() - 1,
  end_date = Sys.Date() - 1)
}
\arguments{
\item{start_date}{first day of data (Date/scalar)}

\item{end_date}{last day of data (Date/scalar)}
}
\description{
Get transactions (<= 7 days only)
}
