% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_order_as_picklist.R
\name{sv_order_as_picklist}
\alias{sv_order_as_picklist}
\title{Convert an order to a picklist}
\usage{
sv_order_as_picklist(saleid)
}
\arguments{
\item{saleid}{SaleId of order to pick}
}
\description{
Convert an order to a picklist
}
