% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_parse_channel.R
\name{sv_parse_channel}
\alias{sv_parse_channel}
\title{Parse Sales Channel from sales data}
\usage{
sv_parse_channel(x)
}
\arguments{
\item{x}{sales data from a call to \code{\link{sv_get_sales}}}
}
\description{
Parse Sales Channel from sales data
}
