% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_parse_order_contents.R
\name{sv_parse_order_contents}
\alias{sv_parse_order_contents}
\title{Get individual items contained within a sale}
\usage{
sv_parse_order_contents(x)
}
\arguments{
\item{x}{data returned from 'sv_get_sales'}
}
\description{
Get individual items contained within a sale
}
