% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_parse_sale_contents.R
\name{sv_parse_sale_contents}
\alias{sv_parse_sale_contents}
\title{Get sale items}
\usage{
sv_parse_sale_contents(x)
}
\arguments{
\item{x}{data returned from 'sv_get_sales'}
}
\description{
Get sale items
}
