% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_pick_order.R
\name{sv_pick_order}
\alias{sv_pick_order}
\title{Auto-Pick an order}
\usage{
sv_pick_order(saleid = NULL, picklist = NULL)
}
\arguments{
\item{saleid}{SaleId of order to pick (character/scalar)}

\item{picklist}{prepared picklist (data.frame)}
}
\description{
Auto-Pick an order
}
