% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_remove_inventory.R
\name{sv_remove_inventory}
\alias{sv_remove_inventory}
\title{Remove inventory}
\usage{
sv_remove_inventory(...)
}
\arguments{
\item{...}{args passed on to \code{\link{sv_update_inventory}}}
}
\description{
Remove inventory
}
