% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_update_po_costs.R
\name{sv_update_po_costs}
\alias{sv_update_po_costs}
\title{Update costs in a skuvault purchase order}
\usage{
sv_update_po_costs(data)
}
\arguments{
\item{data}{input data.frame that includes columns {PoNumber, Sku, Cost}}
}
\description{
Update costs in a skuvault purchase order
}
