% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_update_sent_status.R
\name{sv_update_sent_status}
\alias{sv_update_sent_status}
\title{Update costs in a skuvault purchase order}
\usage{
sv_update_sent_status(
  po_number,
  sent_status = c("NotSent", "Sent", "NeedToResend")
)
}
\arguments{
\item{data}{input data.frame that includes columns {PoNumber, Sku, Cost}}
}
\description{
Update costs in a skuvault purchase order
}
