% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_update_classification.R
\name{sync_sv_classifications}
\alias{sync_sv_classifications}
\title{Update Classification in Skuvault based on inventory contents}
\usage{
sync_sv_classifications()
}
\description{
Update Classification in Skuvault based on inventory contents
}
