% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_add_dropship_inventory.R
\name{sv_add_dropship_inventory}
\alias{sv_add_dropship_inventory}
\title{Add drop-ship inventory to skuvault}
\usage{
sv_add_dropship_inventory(
  skus,
  qty,
  locs = c("STENS", "ROTARY", "POWER", "RBI", "GARDNER")
)
}
\arguments{
\item{skus}{skus to add quantity}

\item{qty}{quantity to add (must be same length as `skus`, or length 1)}

\item{locs}{locations to add (must be same length as `skus`, or length 1)}
}
\description{
Add drop-ship inventory to skuvault
}
