% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_get_transactions.R
\name{sv_get_transactions}
\alias{sv_get_transactions}
\title{Get transactions}
\usage{
sv_get_transactions(
  start_date = Sys.Date() - 14,
  end_date = Sys.Date() - 1,
  sale_id = NULL
)
}
\arguments{
\item{start_date}{first day of data (Date/scalar)}

\item{end_date}{last day of data (Date/scalar)}
}
\description{
Get transactions
}
