% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_update_inventory.R
\name{sv_update_inventory}
\alias{sv_update_inventory}
\title{Update inventory (add/remove)}
\usage{
sv_update_inventory(
  skus,
  qty,
  locs,
  warehouse = "WH1",
  type = c("add", "remove")
)
}
\arguments{
\item{skus}{skus to add quantity}

\item{qty}{quantity to add (must be same length as `skus`, or length 1)}

\item{locs}{locations to add them to (must be same length as `skus`, or length 1)}

\item{warehouse}{warehouse to add them to}

\item{type}{add/remove (scalar/character, must choose one)}
}
\description{
Update inventory (add/remove)
}
